refactor(desktop): split diffviewerdialog into sub-files
